NGSTACK-547 cast return value to integer

+++ bundle/Templating/Twig/Extension/SiteRuntime.php
@@ -120,11 +120,11 @@ class SiteRuntime
         return '/';
     }
 
-    public function calculateReadingTime(string $text): float
+    public function calculateReadingTime(string $text): int
     {
         $wordCount = str_word_count($text);
         $readingTime = floor($wordCount / self::WORDS_PER_MINUTE);
 
-        return $readingTime === false || $readingTime < 1 ? 1 : $readingTime;
+        return $readingTime === false || $readingTime < 1 ? 1 : (int) $readingTime;
     }
 }
